+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-kotlin/runtime/pom.xml
@@ -16,11 +16,7 @@
     <dependencies>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-resteasy-reactive-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkus.resteasy.reactive</groupId>
-            <artifactId>resteasy-reactive-vertx</artifactId>
+            <artifactId>quarkus-resteasy-reactive-server-common</artifactId>
         </dependency>
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
