+++ extensions/resteasy-reactive/quarkus-resteasy-reactive/common-runtime/src/main/java/io/quarkus/resteasy/reactive/server/common/runtime/EndpointInvokerFactory.java
@@ -1,0 +1,10 @@
+package io.quarkus.resteasy.reactive.server.common.runtime;
+
+import java.util.function.Supplier;
+
+import org.jboss.resteasy.reactive.server.spi.EndpointInvoker;
+
+public interface EndpointInvokerFactory {
+
+    Supplier<EndpointInvoker> invoker(String baseName);
+}
