+++ extensions/resteasy-reactive/quarkus-resteasy-reactive/deployment/src/main/java/io/quarkus/resteasy/reactive/server/deployment/QuarkusServerEndpointIndexer.java
@@ -39,6 +39,7 @@ import io.quarkus.gizmo.ClassCreator;
 import io.quarkus.gizmo.MethodCreator;
 import io.quarkus.gizmo.MethodDescriptor;
 import io.quarkus.gizmo.ResultHandle;
+import io.quarkus.resteasy.reactive.server.common.runtime.EndpointInvokerFactory;
 import io.quarkus.resteasy.reactive.server.runtime.ResteasyReactiveRecorder;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.http.HttpServerResponse;
@@ -108,7 +109,7 @@ public class QuarkusServerEndpointIndexer
     protected boolean handleCustomParameter(Map<DotName, AnnotationInstance> anns, ServerIndexedParameter builder,
             Type paramType, boolean field, Map<String, Object> methodContext) {
         methodContext.put(GeneratedClassBuildItem.class.getName(), generatedClassBuildItemBuildProducer);
-        methodContext.put(ResteasyReactiveRecorder.class.getName(), resteasyReactiveRecorder);
+        methodContext.put(EndpointInvokerFactory.class.getName(), resteasyReactiveRecorder);
         return super.handleCustomParameter(anns, builder, paramType, field, methodContext);
     }
