+++ extensions/resteasy-reactive/quarkus-resteasy-reactive/pom.xml
@@ -15,6 +15,8 @@
     <packaging>pom</packaging>
     <modules>
         <module>deployment</module>
+        <module>spi-deployment</module>
+        <module>common-runtime</module>
         <module>runtime</module>
     </modules>
