+++ extensions/resteasy-reactive/quarkus-resteasy-reactive/runtime/pom.xml
@@ -19,8 +19,8 @@
             <artifactId>jboss-jaxrs-api_2.1_spec</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.quarkus.resteasy.reactive</groupId>
-            <artifactId>resteasy-reactive-vertx</artifactId>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-resteasy-reactive-server-common</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
