+++ extensions/resteasy-reactive/quarkus-resteasy-reactive/spi-deployment/src/main/java/io/quarkus/resteasy/reactive/server/spi/MethodScannerBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkus.resteasy.reactive.server.deployment;
+package io.quarkus.resteasy.reactive.server.spi;
 
 import org.jboss.resteasy.reactive.server.processor.scanning.MethodScanner;
